Add files via upload
It takes inputs of log files generated from exonerate and the corresponding nucleodtie fasta files, and output the amino acid sequuences and cds identified by exonerate. In addition, it calculates the p-distance between identified exons and the reference exons and finds the best hits, which can be used for the following exon based sequence alignment.